# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZenFM API
  version: 0.1.0
  description: |
    The local, single-owner ZenFM HTTP API. Browser requests authenticate with
    an opaque `zenfm_session` cookie (`zenfm_session_http` in explicit HTTP
    mode) and send the session-bound value returned
    as `csrfToken` in `X-ZenFM-CSRF` for every mutation. Personal API tokens use
    `Authorization: Bearer`; their restricted scope excludes owner security,
    token management, update, and lifecycle operations.
servers:
  - url: /
paths:
...
  /api/v1/session:
    get:
      operationId: validateSession
      security:
        - sessionCookie: []
      responses:
        '200':
          description: The validated browser session. Validation does not extend its absolute lifetime.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded and a secure session cookie was set.
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HttpOnly, SameSite=Strict session cookie; Secure under HTTPS.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/UniformLoginFailure'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      operationId: logout
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Csrf'
      responses:
        '204':
          description: The current session was revoked and its cookie cleared.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/v1/settings:
    get:
      operationId: getSettings
      security:
        - sessionCookie: []
        - personalToken: []
      responses:
        '200':
          description: Owner preferences and redacted effective server settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      operationId: updatePreferences
      description: Updates owner presentation preferences. Personal tokens cannot change settings.
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/Csrf'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferenceUpdate'
      responses:
        '200':
          description: Updated settings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zenfm_session
      description: HTTPS session cookie. Explicit HTTP mode uses `zenfm_session_http` so a retained Secure cookie cannot block HTTP login.
    personalToken:
      type: http
      scheme: bearer
      bearerFormat: opaque ZenFM personal token
  parameters:
    Csrf:
      in: header
      name: X-ZenFM-CSRF
      required: true
      schema:
        type: string
        minLength: 32
      description: Session-bound CSRF value. Required for browser mutations.
...
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, expired, or revoked credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UniformLoginFailure:
      description: Credentials were not accepted. Account existence is not disclosed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated principal lacks access, setup is incomplete, CSRF failed, or origin was rejected.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Authentication or public-share attempt limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Path:
      type: string
      minLength: 1
      maxLength: 4096
      pattern: '^/'
      examples: [/Books]
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 1024
...
    Session:
      type: object
      additionalProperties: false
      required: [authenticated, setupRequired]
      properties:
        authenticated:
          type: boolean
        setupRequired:
          type: boolean
        csrfToken:
          type: string
          description: Kept only in browser memory by the official client.
        idleExpiresAt:
          type: string
          format: date-time
        absoluteExpiresAt:
          type: string
          format: date-time
...
    PreferenceUpdate:
      type: object
      additionalProperties: false
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        locale:
          type: string
          minLength: 2
          maxLength: 32
        showHidden:
          type: boolean
        clientTimeoutSeconds:
          type: integer
          minimum: 0
          maximum: 86400
        favorites:
          type: array
          description: Ordered file or folder paths relative to Home. New favorites must be existing regular files or directories. Send an empty array to clear favorites.
          items:
            $ref: '#/components/schemas/Path'
        favoriteLabels:
          type: object
          description: Custom display labels keyed by favorite path. Replaces the label map when supplied. Blank or omitted labels use the file or folder name. Labels never change paths.
          additionalProperties:
            type: string
            maxLength: 200
    Settings:
      allOf:
        - $ref: '#/components/schemas/PreferenceUpdate'
        - type: object
          required: [theme, locale, showHidden, clientTimeoutSeconds, advancedMode, root, secureTransport]
          properties:
            favoriteTypes:
              type: object
              readOnly: true
              description: Current type of each accessible favorite path, for choosing its navigation route.
              additionalProperties:
                type: string
                enum: [file, directory]
            advancedMode:
              type: boolean
            root:
              type: string
            secureTransport:
              type: boolean
...
    Problem:
      type: object
      additionalProperties: true
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference